Fix ts import aliases and refactor code to use them
