+++ src/components/Map/index.tsx
@@ -1,1 +1,2 @@
 export * from "./Map";
+export * from "./GeographySelect";
